Fix permission issues when generating reports
The report generator was creating problems because it tried
to create dataset directories in the `processed` directory when
generating reports (and create and write log files). In some
cases, the user might not have a write permission there. This is
fixed now. There is an option to generate dataset directory and
logs when creating datasets, and report generator by default
sets it to False.