Implement mixed beam position method
The beam position is now determined in two stages. In the first
stage, the beam position is computed using the midpoint method.
If the beam vertical position (determined by the midpoint)
falls outside of the panel gap, that means the beam is visible.
In that case, the beam position is recomputed using the maximum
pixel method.

This version also solves the problem with reading corrupted
HDF5 files (when computing the beam position).